requirements.txt
Flask
python-dotenv
requests
python-telegram-bot